# openapi.yaml
openapi: 3.0.3
info:
  title: API Estoque Vitória Cestas
  version: 1.2.0
  description: Endpoints para gestão de itens, fornecedores, autenticação e demais cadastros previstos no banco de dados.
servers:
  - url: http://{host}
    variables:
      host:
        default: localhost:3000
  - url: https://projeto-vitoriacestas-backend.vercel.app
paths:
...
  /api/items:
    get:
      summary: Lista itens cadastrados
      tags: [Itens]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Itens retornados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra novo item
      tags: [Itens]
      security:
        - bearerAuth: []
      description: Disponível para perfis Admin e Gestor.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        '201':
          description: Item criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
        '409':
          description: Código duplicado
  /api/products:
    get:
      summary: Lista produtos cadastrados (alias de /api/items)
      tags: [Itens]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Produtos retornados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra novo produto (alias de /api/items)
      tags: [Itens]
      security:
        - bearerAuth: []
      description: Disponível para perfis Admin e Gestor.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
        '409':
          description: Código duplicado
  /api/upload:
    post:
      summary: Upload de imagem opcional
      tags: [Itens]
      security:
        - bearerAuth: []
      description: Recebe um arquivo de imagem (campo `file`) e retorna metadados básicos do upload. Caso o nome não possa ser extraído do corpo multipart, o cabeçalho opcional `x-filename` será utilizado. Alias disponível também em `/upload`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Upload recebido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload recebido
                  filename:
                    type: string
                    nullable: true
                  mimetype:
                    type: string
                  size:
                    type: integer
        '400':
          description: Nenhum arquivo foi enviado
  /upload:
    post:
      summary: Upload de imagem opcional (alias)
      tags: [Itens]
      security:
        - bearerAuth: []
      description: Alias direto de `/api/upload` para receber um arquivo (campo `file`) e retornar metadados do upload.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Upload recebido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload recebido
                  filename:
                    type: string
                  mimetype:
                    type: string
                  size:
                    type: integer
        '400':
          description: Nenhum arquivo foi enviado
  /api/suppliers:
    get:
      summary: Lista fornecedores cadastrados
      tags: [Fornecedores]
      security:
        - bearerAuth: []
      description: Disponível para Admin, Gestor e Operador.
      responses:
        '200':
          description: Fornecedores retornados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  suppliers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Fornecedor'
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra fornecedor
      tags: [Fornecedores]
      security:
        - bearerAuth: []
      description: Disponível apenas para perfil Admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FornecedorInput'
      responses:
        '201':
          description: Fornecedor criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FornecedorResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
        '409':
          description: CNPJ duplicado
  /suppliers:
    get:
      summary: Lista fornecedores cadastrados (alias)
      tags: [Fornecedores]
      security:
        - bearerAuth: []
      description: Alias sem prefixo /api para compatibilidade com o frontend.
      responses:
        '200':
          description: Fornecedores retornados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  suppliers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Fornecedor'
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
    post:
      summary: Cadastra fornecedor (alias)
      tags: [Fornecedores]
      security:
        - bearerAuth: []
      description: Alias de `/api/suppliers` para criação de fornecedores.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FornecedorInput'
      responses:
        '201':
          description: Fornecedor criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FornecedorResponse'
        '400':
          description: Falha de validação
        '401':
          description: Token ausente ou inválido
        '403':
          description: Permissão insuficiente
        '409':
          description: CNPJ duplicado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ItemInput:
      type: object
      required: [codigo, nome, quantidade, preco]
      properties:
        codigo:
          type: string
        nome:
          type: string
        descricao:
          type: string
        categoria:
          type: string
        quantidade:
          type: integer
          nullable: true
          minimum: 0
          default: 0
        preco:
          type: number
          format: float
          minimum: 0
          nullable: true
          default: 0
        fornecedorId:
          type: integer
          nullable: true
    ItemResponse:
      type: object
      properties:
        message:
          type: string
        item:
          allOf:
            - $ref: '#/components/schemas/ItemInput'
            - type: object
              properties:
                id:
                  type: integer
    Item:
      allOf:
        - $ref: '#/components/schemas/ItemInput'
        - type: object
          properties:
            id:
              type: integer
...
    FornecedorInput:
      type: object
      required: [cnpj, razaoSocial, contato, email, telefone]
      properties:
        cnpj:
          type: string
        razaoSocial:
          type: string
        contato:
          type: string
        email:
          type: string
          format: email
        telefone:
          type: string
        enderecoId:
          type: integer
          nullable: true
    FornecedorResponse:
      type: object
      properties:
        message:
          type: string
        supplier:
          allOf:
            - $ref: '#/components/schemas/FornecedorInput'
            - type: object
              properties:
                id:
                  type: integer
    Fornecedor:
      allOf:
        - $ref: '#/components/schemas/FornecedorInput'
        - type: object
          properties:
            id:
              type: integer